Catch error when type is nil
When geocoding "Porter school, tel aviv" you get a result with an empty type, which was causing a crashing error. This fixed it for me with no side effects so far. Feel free to modify it before merging, this is more like a bug report with a fast solution. Thanks :)